+++ geometrikks/server/core.py
@@ -12,7 +12,7 @@ from litestar.config.compression import CompressionConfig
 from geometrikks.config.settings import Settings, get_settings
 from geometrikks.server import plugins
 from geometrikks.server.exceptions import EXCEPTION_HANDLERS
-from geometrikks.server.lifecycle import on_startup, on_shutdown
+from geometrikks.server.lifecycle import LIFESPAN
 from geometrikks.server.routes import get_route_handlers
 from geometrikks.server.dependencies import create_settings_provider
 
@@ -93,8 +93,7 @@ def create_app(
     app = Litestar(
         debug=settings.debug,
         route_handlers=get_route_handlers(include_auth=not settings.auth_disabled),
-        on_startup=[on_startup],
-        on_shutdown=[on_shutdown],
+        lifespan=list(LIFESPAN),
         plugins=plugins.create_plugins(settings, db_config=db_config),
         dependencies=dependency_map,
         openapi_config=openapi_config,